feat(backup): reload providers on restore
