+++ feature/mobile/settings/src/main/kotlin/com/flixclusive/feature/mobile/settings/screen/root/SettingsViewModel.kt
@@ -1,5 +1,6 @@
 package com.flixclusive.feature.mobile.settings.screen.root
 
+import android.content.Context
 import android.net.Uri
 import androidx.compose.runtime.Stable
 import androidx.datastore.preferences.core.Preferences
@@ -11,6 +12,7 @@ import com.flixclusive.core.datastore.DataStoreManager
 import com.flixclusive.core.datastore.model.system.SystemPreferences
 import com.flixclusive.core.datastore.model.user.BackupOptions
 import com.flixclusive.core.datastore.model.user.UserPreferences
+import com.flixclusive.data.backup.util.BackupUtil.decodeFromUri
 import com.flixclusive.data.database.repository.SearchHistoryRepository
 import com.flixclusive.data.database.session.UserSessionManager
 import com.flixclusive.data.provider.repository.CachedLinksRepository
@@ -19,9 +21,12 @@ import com.flixclusive.data.provider.repository.ProviderRepository
 import com.flixclusive.domain.backup.common.BackupState
 import com.flixclusive.domain.backup.usecase.CreateBackupUseCase
 import com.flixclusive.domain.backup.usecase.RestoreBackupUseCase
+import com.flixclusive.domain.provider.usecase.manage.InitializeProvidersUseCase
 import com.flixclusive.domain.provider.usecase.manage.UnloadProviderUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
+import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
@@ -34,6 +39,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 internal class SettingsViewModel @Inject constructor(
+    @param:ApplicationContext private val context: Context,
     private val userSessionManager: UserSessionManager,
     private val dataStoreManager: DataStoreManager,
     private val searchHistoryRepository: SearchHistoryRepository,
@@ -42,6 +48,7 @@ internal class SettingsViewModel @Inject constructor(
     private val cachedLinksRepository: CachedLinksRepository,
     private val appDispatchers: AppDispatchers,
     private val installedRepoRepository: InstalledRepoRepository,
+    private val initializeProviders: InitializeProvidersUseCase,
     private val _buildConfig: BuildConfigProvider,
     private val createBackupUseCase: CreateBackupUseCase,
     private val restoreBackupUseCase: RestoreBackupUseCase,
@@ -167,10 +174,23 @@ internal class SettingsViewModel @Inject constructor(
         uri: Uri,
         options: BackupOptions = BackupOptions(),
     ): BackupState {
+        val backup = context.decodeFromUri(uri = uri)
+
+        val includeProviders = options.includeProviders && backup.providers.isNotEmpty()
+        if (includeProviders) {
+            deleteProviders()
+        }
+
         return restoreBackupUseCase(
             uri = uri,
             options = options,
         ).first { state -> state is BackupState.Success || state is BackupState.Error }
+            .also {
+                if (includeProviders) {
+                    initializeProviders()
+                        .collect()
+                }
+            }
     }
 
     private suspend fun getCurrentUserId(): String {
